Changes to search for all particles+baryon search
Update to how the initial field halo search is done. If all particles
are searched but a secondary baryon search requested, then search
proceeds differently than it would be for all particle search. It is no
longer simple FOF algorithm. Also code in place but not tested for
keeping the 3dfof envelop.