Number of files per snapshot was not being read correctly from HDF5 snapshots.
In function HDF_get_nfiles() the attribute NumFilesPerSnapshot is read
into either intbuf or longbuf depending on its data type. This buffer is then
indexed using the integer j, which has not been set.

NumFilesPerSnapshot is a scalar in Gadget HDF5 and Illustris snapshots so this
can be fixed by making intbuf and longbuf scalars.