Updates to input to allow user to specify input buffer sizes
updates to input routines and UI to allow user to specify the temporary particle buffer size used in sending information from the read mpi process to those not reading input file. Prior to update input routine can allocate a lot of memory to minimise number of send/recvs if number of mpi processes is large.