Update to KDTree library and interface to library
KDTree now defines a Int_tree_t which can differ from the Int_t type. This allows a user to reduce the memory footprint of the Tree while still using long ints as the Int_t type for items like the particle ID, indexing, etc. This has also changed the interface to certain functions of KDTrees, specifically the FOF routines. FOF routines previously accepted Int_t arrays if memory previously allocated was to be passed to the routines but now must be passed Int_tree_t (which could be the same type as a Int_t)