Initial integration of treefrog
Treefrog depends on nbodlib. Treefrog might also need to be compiled
together with VELOCIraptor to ensure that VR's outputs are compatible
with TF. To this end it was decided that TF will optionally build VR as
well, which is what this patch tries to solve.

The build is still not working perfectly, but is a first step. Next one
is to identify all the nbodylib-specific options down from the VR build
down to nbodylib's, so the latter can be safely included by external
projects like TF and VR.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>