Avoid fitting.h duplication
In MacOS systems filenames are stored case-sensitively, but name lookup and
matching happens case-insensitively. This means that depending on the order in
which the build system adds the -I flags to the compiler either the
src/fitting.h file or the NBodyLib/src/Math/Fitting.h file will match. This, in
turn, means that the incorrect file might be included, leading to compiler
failures due to missing definitions.

A few options to solve this problem could be offered:

 * Ensure that the correct order of -I flags is issued. This might solve the
   particular failure we are seeing now, but most likely will either cause new
   ones, or will still leave the door open for new ones to come.
 * Change the inclusion namespace to contain a hierarchy of at least one folder
   (e.g., #include "Math/Fitting.h"). This would be a very invasive change, but
	would ensure that all include files (not only the one causing trouble now)
   can be uniquely identified.
 * Rename one of the fitting.h files, and adjust the corresponding #include
   statements. This is the solution implemented by this patch because it is the
   less invasive and solves the problem at hand.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>