New read_attribute hdf5 utility functions
These functions take a file (either as a filename string, or as an
H5File object) and a slash-separated attribute name indicating the full
path within the hdf5 file to an attribute, and return its value. Because
attributes can have different types, these functions are templated with
the type of attribute being read.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>