VR_NO_MPI -> VR_MPI
This is to make variable names more standard, given that all the rest are
always defined in a "positive" way and ON by default.

I'm also grouping the VR_MPI option alongside the rest of the MPI-related
options instead of having it under a "dependencies" group. Without MPI in the
group anymore, the remaining options were HDF5 and XDR, and therefore the group
is now called "input format" instead.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>