XDR option is independent from HDF5
This was an early mistake I did during the initial translation of the
Makefile.config.template file, but since then it became clear to me (and
confirmed with Pascal) that these two are completely independent.

As a side-note, XDR support is broken in MacOS 10.11, so users need to manually
opt out from XDR support on such platforms.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>