Explicitly check for GSL in VELOCIraptor
GSL is mostly used in NBodyLib, but VELOCIraptor also uses it directly
in a place or two. We have been getting away with not declaring GSL as a
direct dependency of VELOCIraptor because we are implicitly relying on
NBodyLib to provide us with it, but that might not always be the case.
Let us be nice and add the check so be on the safe side.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>