Adding version checks
NBodyLib, VELOCIraptor and treefrog already have each a VERSION file,
and we were already parsing it. This commit builds on top of that,
exporting the version information for each tool to the parent scope, and
checking that the version of the packages are within certain limits.
These limits of course need to be kept up to date as the software
progresses.

There were some minor problems with some variable names used for storing
the version components which had not been spotted before (because we
were not using this information basically), so those had to be solved.
There was also a subtle problem on how we were reading the VERSION file,
using CMAKE_SOURCE_DIR instead of CMAKE_CURRENT_SOURCE_DIR. This yielded
the wrong version information for NBodyLib and VELOCIraptor when
included from another package.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>