HDF5 1.10.0.1 not to be confused with 1.10.1
What we actually want to differentiate is HDF5 versions >= 1.10.1. We
were previously assuming that anything > 1.10.0 was >= 1.10.1, but since
HDF5 supports even tweak versions this isn't true and the code fails to
compile (i.e., 1.10.0.1 is not >= 1.10.1). A more correct implementation
is then for versions that are NOT < 1.10.1, which yields the correct
response.

The macro name that gets defined in these cases (HDF5_NEWER_THAN_1_10_0)
still reflects this old assumption, but it is a bit of a detail really.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>