Generating a fat velociraptor static library
Instead of linking the VR and nbody static libraries into the final stf
executable we are now generating a fat VR static library that contains
both the VR objects AND the nbody objects. The stf binary then links
only to this single fat static library.

The reason to do this was to simplify the way Swift (and any other
potential users) link against velociraptor. By providing a single, fat
static library with all the necessary code users only need to link
against a single library instead of two.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>