Supporing HDF5 1.8.0+
Because of the usageof H5::H5Location, the code implicitly depended on
HDF5 >= 1.8.12, although the CMakeLists.txt didn't explicitly said so,
so even successful cmake runs resulted on failed builds. By removing its
usage and moving to a templated function solution we can support
versions of HDF5 down to 1.8.0 (I think!), and allows us to build the
software in older platforms, like Travis (and probably many HPC
systems). This is basically the same solution we have in place for
shark.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>