Update to approximative local field density estimation
Updates to code to, at runtime, use either an exact or
approximative near neighbour search to calculate the local
velocity density. There is a MPI+OpenMP bug in approximative search,
and the OpenMP in this portion of this code has been disabled.

Initial tests indicate code roughly produces similar results to exact search
but further testing is required.