Fix for memory corruption when attribute has unexpected type
When reading attributes the code was reading the type of the
attribute from the file and passing this to H5Aread() instead
of the type of the array in memory. This causes a segfault if
the type in the file is not the same.

Modified to pass the type of the array to H5Aread instead.