Fix for segfault when HDF5 attribute has unexpected type (#37)
* Fix for memory corruption when attribute has unexpected type

When reading attributes the code was reading the type of the
attribute from the file and passing this to H5Aread() instead
of the type of the array in memory. This causes a segfault if
the type in the file is not the same.

Modified to pass the type of the array to H5Aread instead.

* Corrections to handling of string attributes on input

Data type in memory is constructed by making a 1 character
C string (H5T_C_S1) and setting its length.