Update to SO radii finding
Previously the code would loop to find the SO radii starting at a point either
from the maximum of SphericalOverdensityMinHaloFac*npartinSOregion or
SphericalOverdensityMinHaloFac*HaloMinSize, which for some haloes can be beyond
the point where the density drops below the desired overdensity. The start
point has now been changed so it always starts the loop from the centre
of the halo.