Update to the I/O so the parallel interface isn't used if MPI_number_of_tasks_per_write==1 and NMPI>1 (#84)
* Update to the I/O so the parallel interface isn't used if MPI_number_of_tasks_per_write==1 and NMPI>1

This update is required since the code would fail if MPI_number_of_tasks_per_write==1 and NMPI>1,
since the code would only create 1 file but would attempt to write to NMPI files causing it to crash.

* Update to SO radii finding

Previously the code would loop to find the SO radii starting at a point either
from the maximum of SphericalOverdensityMinHaloFac*npartinSOregion or
SphericalOverdensityMinHaloFac*HaloMinSize, which for some haloes can be beyond
the point where the density drops below the desired overdensity. The start
point has now been changed so it always starts the loop from the centre
of the halo.