Removed the SINGLEPRECISION flag (#85)
This flag meant that all the calculations would use floats, but it was found
that the loss in precision meant that the code so of the calculations may
produce slightly different results each time it was run. To ensure more
consistency this flag and the use of floats in calculations have been removed.