Add check-before-use for Particle properties
Particles have four optional sets of properties: hydro, star,
black hole, and extra dark matter. Because they are optional, they must
be checked first before using them. If they have not been set by the
time they are used, an error will be caused by dereferencing
uninitialized pointers.

This check is not always in place, and this commit adds some missing
checks. They are in CalculateApertureQuantities and in the different
GetExtra*Properties functions. From what I can tell, until now the code
seemed to have handled this by assuming that the properties are always
set for the particles that are passed to these functions. This is not
true however for particles passed on by the swift-vr invocation
interface.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>