Remove double as return type of ExtraPropFunc
The different functions calculating accumulated metrics over the extra
gas/star/bh properties shared a common signature (made explicitly by the
ExtraPropFunc typedef) signaling that a double value should be returned.
However, in reality none of these functions performed a return. This in
seems to have contributed to the introduction of UB, resulting in
hard-to-debug errors.

This commit modifies the ExtraPropFunc typedef to indicate that
functions adjusting to this type don't need to return a double, and
should return void instead. The signature of all the functions that are
used through this typedef are then adjusted to reflect this change.

Although it looks unrelated, this patch fixes the problem described in
issue #18. This is consistent with the fact that the problem surfaced
only after certain optimization levels were introduced, and with some of
debugging information shown by gdb (although at the time it wasn't
obvious what the problem was).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>