Restore missing object allocations
These objects need to be allocated before they are used, which happens
shortly after in the main body of the subsequent for loop. And because
they were not allocated, using them resulted in segmentation faults.

These allocations were previously part of the code, but they were
removed (apparently by accident) in 842381a.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>