Abort on swift-enable builds later
When the stf executable is run, it immediately exists if the software
was compiled to support integration with swift. This fast failure
however prevents the executable from showing some important information
that can be useful for debugging problems like #23. Moving the fast
exit only a couple of lines down allows for these informational messages
to be printed before the program exits.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>