Remove unused variables
This is but a small set of unused variables visible when compiling the
code with -Wall -Wextra.

I have chosen to fix this file and functions simply to clean up the path
to further debugging the problem described in #37, which stems from the
assignment of zeros to particle densities, which in turn happens in
GetVelocityDensityApproximative.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>