Fix compilation against HDF5 1.12
In HDF5 1.12 H5Oget_info is a macro that points either to H5Oget_info1
or H5Oget_info3, which have different signatures. This patch checks the
HDF5 version we are compiling against and uses the correct signature to
invoke this function. The same applies to H5Oget_info_by_name{,3}.

This fix is based on pelahi@72f90bf but is not a direct cherry-pick, as
the original contained more changes than what I'd like to bring, plus it
doesn't correctly invoke (the new) H5Oget_info with correct values.

This addresses #42.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>