Fix macro name typo to enable HDF5 compression
This was noted in #45. In the past the macro was defined (with the typo)
within the same file, but later on its definition was moved out to the
cmake infrastructure, where the correct spelling was used, hence the
incorrectly-spelled #ifdef block now always evaluate to false.

Since the names used to define and check the macro matched in the past,
I suspect the code within the #ifdef block was indeed exercised at some
point, and that it is correct (and at least it doesn't look obviously
incorrect). Hopefully this is still the case, and we won't be
introducing a new bug by allowing this code to compile.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>