Bulk improvement on memory usage reporting
Instead of manually building the file/line/function placement where the
memory reporting is taking place, we now provide a macro that provides this
information automatically for us, and we then go and construct the
corresponding message in a nice, uniform way. The output is not the same
as it used to be, but actually a bit improved (but could still be
simplified).

I also took the chance of changing several logging statements to use the
new LOG infrastructure, plus timing of code that could use the new Timer
class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>