Added SO calculations for hot gas and fixed related bugs
- For the total subhalo mass in hot gas and related properties the calculation was incomplete as it
was missing the large halos. This has been fixed.

- Added spherical overdensity calculations for hot gas in arbitrary radial bins and normalised to
an arbitrary input overdensity. Although this is working, it appears like there are still problems
with the calculations as the total SO_500_rhocrit that I compute is not the same as the current outputs.
I suspect this is because the function I used to compute in VR is outdated and need to use a different one.