De-duplicate InitVelociraptor* functions
The body of the two existing functions was almost identical, with that
from InitVelociraptorExtra *not* initialising a couple of things, which
looks like a bug. Their major difference was the Options variable that
they worked on, so instead of duplicating the code it's much easier to
reuse it and pass the corresponding Options object as an argument.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>