Homogenised Rhalfmass calculation
- I'm not using interpolation to compute all half-mass radii, and do that by
looping over particles that are ordered in radii.

- The previous calculation for Rhalfmass_star was wrong while the aperture r50 were correct
(this was determined by comparing the VR catalogues with those of SUBFIND in EAGLE). It seems this was because
the particle rc was defined differently in Rhalfmass_star than in the aperture calculation, and because
Rhalfmass_star wasn't using interpolation. I have now updated the Rhalfmass_* calculations to follow
the aperture ones.