Remove old, commented macro definition
The USEHDFCOMPRESSOIN (sic) macro apparently was originally defined
automatically depending on the HDF5 version and on the value of the
USEPARALLELHDF macro. This changed at some point in time, when the macro
definition was moved to the cmake machinery, giving users an option to
opt in or out of this feature. Thus, the old commented code doesn't
serve any purpose, while also featuring the old, typo'd macro name.

This commit removes the commented code altogether to remove further
confusion.

This addresses #55.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>