Reserve correct amount of memory
The MPI sendrecv operations happening a few lines below require
proprecvbuff to be the same size of the corresponding propsendbuff; that
is, numrecv * numextrafields. This became clear when running the program
under valgrind, and after inspecting it under a debugger.

This commit fixes the immediate problem. A separate one will be added to
add assertions on the sizes of the send buffers to add clarity to the
code.

This addresses #54.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>