Bug fix for non-mpi runs
This fixes a bug that was introduced at some point for non-mpi VR. There was the incorrect addressing of an array using a index i rather than the correct Part[i].GetID() index. This has now been fixed. MPI did not have this error as i=Part[i].GetID() for the mpi runs at this point.

Conflicts:
	src/search.cxx