Small change to SO calculation to follow so-gasstar-masses
I added a small modification to the SO calculations to follow
what Pascal has in his branch so-gasstar-masses but keeping the
interpolation in log10 space.

With this I find that the M200crit halo MF of EAGLE between VELOCIraptor
and SUBFIND are in perfect agreement down to ~32 particles, which is the limit
imposed in SUBFIND to find subhalos.