Include Mass_tot_incl dataset declaration unconditionally
There was a mismatch between the declaration of the datasets outputted
by VR and the actual writing of data at the end of the program. After
some analysis it became clear that the declaration was wrong, as it
contained the Mass_tot_incl dataset declaration within conditions that
shouldn't have been there.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>