Fix incorrect parsing logic
This code, as previously written, entered into the "if" body even when
no separator ("=") was found. By good chance this didn't cause any
issues: because of how std::string::substr works the two calls issued
using the position value "j" ended up not throwing exceptions, although
both the parameter name and value would end up being equal to the full
line. This in turn means none of the subsequent "if" statements that set
different options are invoked.

Although this was probably never seen in the wild because configuration
files are seldom incorrectly written, a warning was generated by
different compilers (depending on version and warning flags). Since the
code is quite warning-full we never paid much attention to it until now.

The new code now properly checks that the result of the find call is
valid before using it.

This addresses #58.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>