Guard access to typeparts
The typeparts vector is only meaningful when compiling with either gas,
srart, BH or high resolution flags, and therefore all its usages should
be guarded like that. This had been the case before, but in 9af51df we
inadvertently broke this while trying to merge the changes from bf9d9a1.

This change addresses #63.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>