Avoid memory leak on MPISetFilesRead calls
Almost all calls to MPISetFilesRead resulted on small memory leaks
because callers were allocating the array that was passed as an argument
(as a pointer reference), which then internally was allocated again,
creating a leak even if the caller was careful of deallocating what they
had allocated.

To make things simpler in the long run I replaced all these arrays by
std::vector objects, ensuring their memory is automatically managed
through RAII, and cleaning up the interface of MPISetFilesRead a bit.

This addresses #65, but also similar memory leaks that were hidden in
the code handling the other input formats (RAMSES, Gadget, NChilada).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>