Declare massval within private OpenMP context
The massval variable was previously declared outside the confines of the
OpenMP parallel context, but it wasn't declared as a private variable
during the omp pragma. This led to race conditions when using its values
from different threads, which eventually resulted on invalid SO mass
quantities for almost all the values calculated within GetSOMasses.

This commit removes the declaration of massval at the beginning of the
GetSOMasses function, and moves it to the innermost context possible,
ensuring its value is not shared across threads, preventing race
conditions, and therefore fixing the values for the SO mass quantities
when running with OpenMP.

This addresses #62.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>