Move non-positive particle density errors up
After some experiments I was able to reproduce the non-positive density
errors earlier on in the code, right after all particles have their
densities calculated. I thus decided to refactor the creation of the
error message out of its previous place (localbgcomp.cxx) and move it
into an exception that can be easily thrown from various places in the
code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>