Report structure opening centrally
Instead of manually logging the opening of groups and datasets, we can
simply log this inside the functions that are doing the group/dataset
opening, and thus centralise this logic.

In the case of datasets we need to do some small invocations to get the
name of the containing group, but nothing too difficult.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>