Simplify Options class
The Options class' members were all first declared, then initialised in
the body of the default constructor. This is firstly against the idea of
initialising members in the member initialisation list, but moreover it
required having to keep two lists of members in the code: the one
declaring them, and the one initialising them.

This commit brings together member declaration and initialisation, which
has been available since C++11. While doing so I found a few members
that were:

 * Not initialised (see #79 for example).
 * Initialised later on deep within VR (so probably shouldn't belong in
   this class in the first place).
 * Uninitialised and later on only written to (and never read).
 * Not used at all.

I've removed the latter two sets of members, and added initialisation
for the first two.

I've also taken the opportunity to remove all the rest of the member
functions, including the constructor, as they were unnecessary.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>