Remove unnecessary, problematic vector initialisation
The SO_mass_interloper vector gets already filled with zeros when
resized, so there's no need to re-initialise it to zero again. Moreover,
this code didn't take into account the same condition used for resizing
the vector, and therefore it crashed when assigning values into an empty
vector.

This is the cause of the segmentation fault reported in #78.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>