Avoid warning due to function not returning values
The mycNFW_fdf function declares it returns a double, but its
implementation doesn't return anything, raising a warning with most
compilers. This commit changes the function to return void. The function
doesn't seem to be used anywhere in the code though, and while in
principle it could have been removed I decided to keep it for future
reference in case it's needed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>