Use vector for automatic memory management
The cellnodeids Options member variable was previously defined as a raw
pointer, which was allocated using new[] but never deallocated via
delete[], thus leading to a (small) leak. Using a vector for automatic
memory management is a better alternative and offers the same indexing
operator used throughout the code.

In the case of the swift interface, we now create a new vector every
time with the contents of the cell_node_ids pointer, which should have
length s.numcells.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>