Rework bits of H5OuptutFile
The H5OutputFile contained some pieces of logic that were very hard to
follow easily around the create/append/close methods. Also, these
methods were defined in the header file itself, but this only adds
complexity to the header file and adds to the compilation time.

This commit both re-implements some of the H5OutputFile class methods
(the create/append/close public methods, plus associated private
member), and moves the actual definition of the methods to a new
h5_output_file.cxx module, leaving the header thinner and easier to
follow.

The implementation of H5OutputFile uses the FileExists function. This
function is only declared in proto.h, which is a single, big pool of
file definitions. We have been trying to move away from that and into a
per-module header file; therefore we used this opportunity to create a
new io.h file and move the declaration of FileExists into it. This
required including the new header in a few modules, all of which are
included in this commit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>