Fix compilation for gas-only builds
There were a few problems with the computation of newly introduced
group-level properties, which caused gas-only builds to break.  In
particular, the M_gas_nsf_incl and M_gas_sf_incl properties of PropData
should only exist when gas AND stars are enabled, while the code
previously defined them when gas was ON and regardless of whether stars
where on or off. Defining them within the correct compilation boundaries
made it clear which other changes were required across the rest of the
code base.

This commit takes care of these problems, adding the #ifdef macros where
needed to avoid compilation issues. While the #ifdef inferno has grown,
it's beyond the scope of this commit to take care of this.

This issue was originally reported in #70, and then again in #91.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>