Improve how compile options/definitions are set
Firstly, instead of setting the compile definitions in the velociraptor
target as PRIVATE and then having to set them again on the executable,
we set them as private and let them trickle down to all targets that
link against velociraptor.

Similarly, the compile options are now set via target_compile_options
publicly instead of manually setting the target property (which won't
trickle down to linking targets).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>