Remove redundant parameters from write_dataset
The only "flag" parameter from the write_dataset family of functions
that was being actively used was flag_parallel; the rest were always
left to their default value (true) and therefore served only to pollute
the interface.

This commit removes these parameters, defining them as local, fixed
variables in the final routine that performs writing. Later on we can
simplify the code in that function based on these fixed values.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>