Cleanup and move rest of write_dataset functions
Now that the main write_dataset_nd function has been cleaned up and
moved to the h5_output_file we can move the rest of the functions that
depend on it. During this move I removed some small duplicate checks,
adjusted indentation, and added asserts for otherwise unnecessary
arguments.

The same treatment was given to write_dataset templated functions, which
must remain in the header.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>