Write HDF5 datasets in 2GB chunks
HDF5 doesn't properly support writing >= 2GB of data in a single I/O
operations. Even though it was reported that 1.10.2 fixed this
limitation, we have seen several situations in which this limit breaks
our code in strange ways.

This commit changes the actual writing of the dataset. Instead of
writing the given data on a single H5Dwrite call, we perform writes as
big as 2 GB at a time. Chunking is done only in the first dimension,
which is also the dimension that is written in parallel from multiple
MPI ranks.

I also took the opportunity to further adjust some variable names and
simplify the code in the write_dataset_nd function.

This is the main change required to remove the problems we were reported
in #88.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>