Size reception buffer on MPI_Bcast correctly
The previous version of this code was both repeated and buggy. The bug
was in the assumption that the reception buffer (sized after the
current, local size of the strings being broadcasted) could hold
whatever rank 0 sent, but that doesn't always hold: rank 0 might be
trying to send a longer string, in which case the code would run into a
buffer overflow and potentially crash.

This commit adds an extra MPI_Bcast call to first send the size of the
string that will be sent in the next step. I also took the opportunity
to remove all duplication and instead define this in a neat function
that is then called appropriately.

This solution might not be the best MPI-wise given that we are sending a
vector of strings, but is sufficient.

This was reported in #100.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>