Use correct types for sending vector sizes
The fix introduced in 1fa1e1e for #87 had itself a bug: we first stored
the size of the vector to be communicated as an int, and then declared
MPI we had an MPI_Int_t, which maps differently to MPI_LONG_LONG_INT or
MPI_INT depending on the LONGINT macro being defined or not. This
therefore doesn't work universally, leading to buffer overflows.

This commit declares the size variables as long long, and uses the
MPI_LONG_LONG_INT type explicitly during the MPI call. This makes the
communication safe regardless of the LONGINT macro being defined or not.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>