Avoid copying Coordinate argument
The xpos argument in MarkCells doesn't need to be copied from the caller
site to the callee, a constant reference is enough.

Moreover, I found there was a misalignment in the signature of
MarkCells, with the declaration missing a Double *r2val argument. Since
the function is local to unbund.cxx I simply removed it from proto.h.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>