Fix code to realise its original intention
This piece of code clearly was meant to cap the Ncellfac value given in
the options so that, when multiplied by 8, didn't go above
MAXCELLFRACTION. Since it was missing a call to std::min this turned out
to be simply a ()-delimited, comma-separate list of values, which
evaluates to the last of them -- in other words, nf always took the
value MAXCELLFRACTION, regardless of the Ncellfac option.

This commit fixes the implementation of this operation, adding the
missing call to std::min to achieve the desired purpose. Since one of
the arguments and the result are both of type Double_t we need to make
sure the constants are also of the same type; otherwise std::min might
complain.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>