Restructure code to avoid duplications
When writing the first datasets of the SO properties hdf5 output file,
there were two branches of the code that wrote them in exactly the same
way. Instead of duplicating that code we can instead structure the logic
flow to have a single block of code serving the two original cases.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>