Add missing broadcast
The broadcasting of the opt.bh_internalprop_output_names vector was
something that originally existed in the codebase, but was accidentally
removed in 4cdbd52 while fixing #100. This missing broadcast meant that
different ranks had different values in this vector, which ultimately
affects the HDF5 dataset count and names that are written by each rank.
When using parallel HDF5 writing this led to misaligned collective
calls, and therefore to the problems described in #102, which is the
issue this commit fixes.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>