Fix incorrect assertion
When writing HDF5 files in parallel, attribute writing must happen in
parallel. Sometimes it's not possible to bypass this limitation, and
therefore attribute writing must be done by a single rank even on
MPI-able scenarios.

The H5OutputFile class has been designed with this usage in mind. While
refactoring the code as part of #88, I added some assertions to make
sure this held true from the caller site (see e703f6f). This assertion
is faulty however, as it only applies when compiling against MPI-enabled
HDF5 libraries, which is what this commit fixes.

This faulty assertion was reported in #113.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>