Mark particles as visited as soon as they are
There was a delay between particles being visited, and being marked as
such (the loop for marking particles as visited came right after the
loop using particles). This worked well when individual calls to
SearchBallPos returned lists of indices without repetitions; if a list
of indices contains repetitions though (and the SearchBallPos algorithm
doesn't prevent this) then the nexport variable could be increased
beyond the extents of the PartDataIn allocation and cause problems,
which is exactly what happens in #73.

This commit moves the marking of visited particles to be within the same
loop where that particle is visited, right after the check.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>