Avoid using shared variable across parallel OpenMP regions
When introduced, the temp variable followed the convention used for the
rest of the GetProperties local variables: it was declared at the top of
the function, then used whenever required. This variable however needed
to be private within each parallel OpenMP region, and therefore also
required being added to the "private" section of the required omp
pragmas.

This commit changes how temp is declared, moving it from the top of the
function to the innermost scope possible. This removes the need to keep
track of it in the private/shared lists of omp pragmas, making the code
less prone in the future.

This addresses the problem reported in #117.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>