Remove unnecessary, unprotected write to empty vector
Just like in 53c0289, the SO_angularmomentum vector only has a certain
size in some situations, and that check wasn't performed in this code
that was setting its values to zero.

Moreover, the values don't need to be initialised to zero as they
already are set to zero when the vector is sized in PropData.Allocate().

This addresses the problem reported in #116.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>