Split memory usage collection and reporting
It will be useful in some places of the code to measure memory usage but
not necessarily to display it as a normal "report" in the logs. Thus,
splitting metrics collection from reporting will help us reuse the
collection bits in other places of the code.

While doing this I also took the chance of simplifying how these memory
collection metrics are performed. For starters, we don't need to know
about that many fields, as we're usually only interested on VM/RSS
curernt/peak usage. Additionally we can use a single file as the source
of the information, which should help speeding up this data collection
as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>