Finish VR graciously on uncaught exceptions
VR's main function didn't have a high-level try/catch block to handle
uncaught exceptions rising from the code, leaving the behavior of the
program to the fate of whatever default the C++ runtime library
provides.

This commit adds such a block. When an uncaught exception is received,
an error is logged, and then either an MPI_Abort is issued in
COMM_WORLD, or a simpler exit is invoked on non-MPI builds.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>