Update to HDF5 interface based on commits from icrar fork
These are fairly simple but do the work. In time we might want to add
more bits of code like this for other containers though.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Add verbosity flag to H5OutputFile class

This will allow us to set the flag on a per-file basis whenever we want
to print more information about the output of a particular file.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Output written dataset names and dimensionality

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Simplify safe_hdf5 function usage

The previous incarnation of this function required callers to specify
the function return type, but safe_hdf5 already has all the information
required to automatically calculate this.

This new version of safe_hdf5 automatically calculates the return type,
and also prints the hdf5 function name in the error message. To do so
without much friction safe_hdf5 is actually a macro that feeds
_safe_hdf5, the actual function.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Use safe_hdf5 in write_dataset_nd

Most of the hdf5 operations performed within write_dataset_nd had their
error code or result not checked, and therefore we could be failing
silently before the actual data writing happens.

This commit adds usage of safe_hdf5 in the HDF5 operations occurring in
write_dataset_nd, thus potentially alerting us of otherwise silent
errors.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Rework bits of H5OuptutFile

The H5OutputFile contained some pieces of logic that were very hard to
follow easily around the create/append/close methods. Also, these
methods were defined in the header file itself, but this only adds
complexity to the header file and adds to the compilation time.

This commit both re-implements some of the H5OutputFile class methods
(the create/append/close public methods, plus associated private
member), and moves the actual definition of the methods to a new
h5_output_file.cxx module, leaving the header thinner and easier to
follow.

The implementation of H5OutputFile uses the FileExists function. This
function is only declared in proto.h, which is a single, big pool of
file definitions. We have been trying to move away from that and into a
per-module header file; therefore we used this opportunity to create a
new io.h file and move the declaration of FileExists into it. This
required including the new header in a few modules, all of which are
included in this commit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Simplify FileExist implementation

There were a few unused or unnecessary variables, now the implementation
is a two-liner.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Move safe_hdf5 into new h5_utils module

Having a h5_utils header+implementation module is in general probably a
good idea, as it will help moving some of the more general utilities out
of hdfitems.h, which currently has defined loads of code inline.

Add new dataspace_information function

This will allow us to inspect more closely the datasets we have opened,
either for reading or for writing.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Inspect datasets before they are written

This will hopefully help us identify what's going on.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Remove redundant parameters from write_dataset

The only "flag" parameter from the write_dataset family of functions
that was being actively used was flag_parallel; the rest were always
left to their default value (true) and therefore served only to pollute
the interface.

This commit removes these parameters, defining them as local, fixed
variables in the final routine that performs writing. Later on we can
simplify the code in that function based on these fixed values.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Cleanup and move main write_dataset_nd function

The implementation of the main write_dataset_nd function was slightly
confusing, as it contained some code duplication and too many #ifdef
macros spread around the code. Additionally, it still used the different
"flag" variables that we now know are always true.

This commit re-implements this routine. The overall structure is still
the same, and the logic remains untouched. However variables have been
renamed for clarity, common code has been factored out into reusable
functions, #ifdef macro has been greatly reduced, and the code has been
simplified after removing all usage of the "flags" variables.

Since this is a big change, I'm taking the opportunity to move the
implementation of this function down to the new h5_output_file.cxx file,
thus avoiding re-compiling it each time we include the hdfitems.h
header.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Cleanup and move rest of write_dataset functions

Now that the main write_dataset_nd function has been cleaned up and
moved to the h5_output_file we can move the rest of the functions that
depend on it. During this move I removed some small duplicate checks,
adjusted indentation, and added asserts for otherwise unnecessary
arguments.

The same treatment was given to write_dataset templated functions, which
must remain in the header.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Unify write_attribute method implementations

There were two very similar implementations of this functionality, one
for plain scalar values and another one for string, but at the end they
were very much the same.

This commit puts the common bits of both implementations in a private
method, then adjusts the two existing public methods to invoke the
private one with the correct arguments.

In addition to this unification, the version of write_attribute taking a
string value was failing to close the new data type it created, leading
to object identifier leaks in HDF5. This has been now fixed.

Finally, scalar attributes can (and are, in VR) only be written in
serial. A new assertion to this effect has been added to the code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Write HDF5 datasets in 2GB chunks

HDF5 doesn't properly support writing >= 2GB of data in a single I/O
operations. Even though it was reported that 1.10.2 fixed this
limitation, we have seen several situations in which this limit breaks
our code in strange ways.

This commit changes the actual writing of the dataset. Instead of
writing the given data on a single H5Dwrite call, we perform writes as
big as 2 GB at a time. Chunking is done only in the first dimension,
which is also the dimension that is written in parallel from multiple
MPI ranks.

I also took the opportunity to further adjust some variable names and
simplify the code in the write_dataset_nd function.

This is the main change required to remove the problems we were reported
in #88.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Close objects in exact reverse order

This might not be a problem in reality, but better safe than sorry.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Add test program for H5OutputFile

This in general is a useful tool to have (and even the beginning of a
unit test suite...), and in particular it helps us reproduce #88 much
faster.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Fix incorrect assertion

When writing HDF5 files in parallel, attribute writing must happen in
parallel. Sometimes it's not possible to bypass this limitation, and
therefore attribute writing must be done by a single rank even on
MPI-able scenarios.

The H5OutputFile class has been designed with this usage in mind. While
refactoring the code as part of #88, I added some assertions to make
sure this held true from the caller site (see e703f6f). This assertion
is faulty however, as it only applies when compiling against MPI-enabled
HDF5 libraries, which is what this commit fixes.

This faulty assertion was reported in #113.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>